test: 피드백적용 - DeleteHistoryTest에서 model 테스트와 repository 테스트를 분리
